requirements.txt
anthropic
langchain>=0.2.0
langchain-core>=0.2.0
langchain-anthropic
langchain-classic
langchain-community
langchain-openai
langgraph
wikipedia
python-dotenv
pydantic
fastapi
uvicorn
joblib
scikit-learn
xgboost
streamlit
pandas
numpy
duckduckgo-search
ddgs
langserve[all]
sse-starlette
uvicorn
langserve
ollama
langchain-chroma 
chromadb
langchain-huggingface
sentence-transformers
pypdf
langchain